restruct files
