+++ routes/index.js
@@ -1,18 +1,10 @@
 
 const express = require('express');
-const bodyParser = require('body-parser');
-const cors = require('cors');
-const DbUtils = require('./dboperation');
-const Member = require('./member');
+const DbUtils = require('../models/dboperation');
+const Member = require('../models/member');
 const { request } = require('http');
-const app = express();
 const router = express.Router();
 
--
-app.use(bodyParser.urlencoded({extended:true}));
-app.use(bodyParser.json());
-app.use(cors());
-app.use('/api',router);
 
 router.use((req,res,next)=>{
   console.log('middleware');
@@ -25,7 +17,6 @@ router.route('/members').get((req,res)=>{
   })
   
 })
-// http://localhost:8090/api/members postman test - get 
 
 router.route('/members/:id').get((req,res)=>{
   DbUtils.getMember(req.params.id).then(result =>{
@@ -42,10 +33,4 @@ router.route('/members').post((req,res)=>{
   
 })
 
-const port = process.env.PORT || 8090;
-app.listen(port);
-console.log('Member API is running at '+port);
-
-
-
-
+module.exports = router;
